config: add new mouse binding ‘launch-url’, bind to alt+BTN_LEFT
This adds a new mouse binding, launch-url (bound to Alt+BTN_LEFT by
default). When executed, the URL under the mouse pointer is
launched (with url.launcher, as usual).

This is done by calling urls_collect(), to get all currently visible
URLs (both auto-detected and OSC-8 ones), and then iterating them all
until we find one that spans the cell the mouse pointer is on.