wip: support *all* modifier names
This is done by no longer validating/translating modifier names to
booleans for a pre-configured set of modifiers (ctrl, shift, alt,
super).

Instead, we keep the modifier *names* in a list, in the key binding
struct.

When a keymap is loaded, and we "convert" the key binding, _then_ we
do modifier translation. For invalid modifier names, we print an
error, and then ignore it. I.e. we no longer fail to load a config due
to invalid modifier names.

Left to do: we still build the "significant" modifier list (used when
matching key bindings) using the same set of pre-configured modifiers;
ctrl/shift/alt/super for matching key bindings, and the additional
modifiers caps-lock and num-lock for the kitty keyboard protocol.

Not quite sure how to handle this at the moment.